requirements.txt
streamlit
google-genai
scikit-learn
numpy
pypdf2
python-docx
python-dotenv